+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/ImageDao.java
@@ -3,11 +3,11 @@ package ar.edu.itba.paw.persistence;
 import ar.edu.itba.paw.models.*;
 
 public interface ImageDao {
-    Image getImage(long id);
+    Image getImage(int id);
 
-    long uploadPop(byte[] bin, String fileName, Booking booking);
+    int uploadPop(byte[] bin, String fileName, Booking booking);
 
-    long uploadVehicleImage(byte[] bin, String fileName, Vehicle vehicle);
+    int uploadVehicleImage(byte[] bin, String fileName, Vehicle vehicle);
 
-    long uploadPfp(byte[] bin, String fileName, User user);
+    int uploadPfp(byte[] bin, String fileName, User user);
 }
