+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/UserDao.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface UserDao<T extends User> {
     boolean mailExists(String mail);
 
-    Optional<T> findById(long id);
+    Optional<T> findById(int id);
 
     Optional<T> findByUsername(String username);
